pm: Ensure PD isn't busy before modifying module state
When disabling a module or power domain, we optimize
operation time by not waiting for the operation to complete
before returning success. In the case of a device being
quickly turned off and back on this can lead to attempting
to turn a module on that is still in the process of turning off.

The solution is to verify previous transitions have completed
before modifying the state register.

Fixes: SYSFW-3182 #integrate-and-build

Signed-off-by: Russ Dill <russ.dill@ti.com>