pm: clk: Reduce search space for clk_div
When trying to find an optimum frequency for a divider that
has a parent frequency that can be programmed, the search
space can be unnecessarily large, especially if the parent is
a PLL and the min/max range is 0 and UINT_MAX.

Reduce the search space by two methods. The simplest
method is to stop the search if the target frequency is
reached.

The second is to bring the min/max values passed to the
parent clock in tighter as more optimum frequencies are
found. The new clock frequency must be better than the
current best frequency by at least 1Hz. So for instance if
the original query is (min/target/max), 0/500/UINT_MAX
and the first round finds a possible frequency of 505Hz
the min/max is updated to 496/504. Any new frequency
found must be within 4Hz of the target.

Fixes: SYSFW-3836 #integrate-and-build

Signed-off-by: Russ Dill <russ.dill@ti.com>