common: mmr_lock Added logic inside the PM functions for mmr_lock/unlock
Cleaned up the PM logic to have the mmr_lock and mmr_unlock functions being
called from the PM functions for which it is required. The other functions
as a result are not impacted.

Additional logic is added to mmr_lock functions to make sure they do not
end up locking the MMRs when there is a context switch and another PM
function is called while processing another PM function.

Fixed: SYSFW-3861

Signed-off-by: Piyali Goswami <piyali_g@ti.com>