pm: Add support for powered-on SWRstDisable state
This adds PSC driver support for setting LPSC modules to
the SWRstDisable state while remaining powered up. This
state is entered if the module is set to disabled,
retention is enabled, and the MRST reset is active.

A LPSC struct member is added to track if this state is
active in order to properly keep track of the necessary
powerdomain reference count.

Fixes: SYSFW-3390 #integrate-and-build

Signed-off-by: Russ Dill <russ.dill@ti.com>