common/am6x/mmr_lock.c: Fix all MISRA-C issues
The most of the changes are quite trivial and easy to understand. However,
the MISRA.INCR_DECR.SIDEEFF.2012 issues fixed by changing -- and ++ operator
use to simple = +1 and -1 is fixing an obvious false positive. But since
getting a Deviation Report trough is so tedious job and there is only couple
of false positives of this kind, it is easier just to get rid of ++ and --
usage all together. Anyway, when following MISRA-C rules, there really is
no benefit in using increment or decrement operators anywhere.

Fixes: SYSFW-1510

Signed-off-by: Jyri Sarha <jsarha@ti.com>