rm: core: Set resource response first range even if num is zero
In some cases RM board configurations may have a
zero length range assigned to a host.  When this
occurred the entry was not returned as the first
entry in the get resource range response.  Update
the get_range function to return the zero length
range first if it's listed first in the board
configuration.

Fixes: SYSFW-4160

Signed-off-by: Justin Sobota <jsobota@ti.com>