pm: clock: Add CLK_DATA_FLAG_NO_HW_REINIT flag
Introduce a CLK_DATA flag, provided by autogen, to prevent PLLs and dividers
from being re-initialized in HW if either:

* A PLL is locked before we try to configure it.
* A divider is programmed to something different than the
 reset value.

With split FW architecture now there is potential for clocks to be
configured by R5 bootloader to frequencies that do not match OFC, and
then later when pm_init runs as part of sciserver start, they get
reconfigured back to OFC state. For now, it is acceptable that this
happens in most cases as handover of devices is not something that is
officially supported, however certain devices are extremely sensitive to
a clock configuration change, particularly DDR.

This functionality will allow clocks known to have issues to be flagged
and then prevent their own re-init to improve system stability.

Fixes: SYSFW-4174 #integrate-and-build

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>