pm: soc: Sync with latest data for CLK_DATA_FLAG_NO_HW_REINIT flag
For both j721e and j7200 we must set the CLK_DATA_FLAG_NO_HW_REINIT for
Main PLL 12 and the associated hsdiv0 so that the DDR PLL is not
reconfigured if it is not locked.

Import from autogen to get this data, while also getting the empty flags
value for all other 16fft PLLs.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>