pm/lib/ilog.c: Fix couple of MISRA.ETYPE.INAPPR.OPERAND.BINOP.2012
The choice if signed return value is a bit unfortunate here. Logarithms
never return negative values if the input value is a positive
integer. With zero the returned value should be an infinite negative
value, and I do not think it is needed. Anyway I do that separately
later if I have time for it.

Fixes: SYSFW-1510

Signed-off-by: Jyri Sarha <jsarha@ti.com>