lpm: Introduce basic low power stub
Introduce the basic foundation for a low power stub which is intended
to be a small independent block of code used for specific low power
tasks while most of the system is powered off.

This patch introduces the functional entry point of a DM stub, called
dm_stub_entry, that will contain all of the stub functionality.

Fixes: SYSFW-3941 #integrate-and-build

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>