rm: ra: fix identification of RA instance for UDMAP
More than one RA inst type can be included in a single
NAVSS when doing searches for the RA inst for UDMAP
pairing.  Selecting the proper RA in the NAVSS based
on the NAVSS ID alone is no longer sufficient.
Added a check for the ring_type based on UDMAP ch_type.
This helps to clarify which RA instance in the NAVSS
should be the one paired with the UDMAP instance.

Fixes: SYSFW-4287
Fixes: SYSFW-4732

Signed-off-by: Jonathan Bergsagel <jbergsagel@ti.com>