lib: trace_protocol: Introduce LPM SEQ trace action and values
Add a TRACE_PM_ACTION_LPM_SEQUENCE action to allow outputting trace
"checkpoints" throughout the low power sequence. Define corresponding
values for this action so that the trace sequence can output these
checkpoints to determine where in the suspend sequence things progress.

This action will follow the standard FAIL flagging as well and will be
OR'd with TRACE_PM_ACTION_FAIL for checkpoints that are reached but for
which the corresponding action has failed.

Finally, only use 0x3fff of the `value` portion of this action to along
a single byte at the end to be used for variable data with each step
value, if needed.

Fixes: SYSFW-4698 #integrate-and-build

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>