pm: clk_mux: set_parent: consider new parent divider also
Currently the set_parent operation doesn't change the parent if the
new parent and current parent is same. But there are scenarios where
switching is requested to an option which has same parent clock as before, but
the divider value is different. This change considers the current parent and
new parent divider value also before deciding to not switch parent.

Example of such scenario is AM64 MCU M4FSS_CLKSEL.

Fixes: SYSFW-4717

Signed-off-by: Vishal Mahaveer <vishalm@ti.com>