pm: clk: Add top level clk LPM functions
To allow saving and restoring of all clocks with drivers that implement
suspend and resume handlers, implement top level calls, clks_suspend and
clks_resume, to loop through all clocks in the device and call relevant
LPM handler.

The suspend path does not currently have any dependency on parent/child
relationship but the resume path currently ensures that any clock with a
parent that is still suspended does NOT get resumed until the parent has
been resumed. This is critical to ensure parent muxes are set up with a
valid clock source.

Fixes: SYSFW-4820 #integrate-and-build

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>