pm: clk: mux: Refactor clk_mux_get_parent
Currently the clk_mux_get_parent function returns a struct clk_parent
but the clk_mux_set_parent function takes an int value representing
which parent to set. Because of these asymmetric get/set calls, a user
can not call get parent and use that same value with set parent.

However, the equivalent int value used in set_parent is just used to
lookup the return value used for get_parent, so this can easily be
refactored to provide a clk_mux_get_parent_value function that is
symmetric with clk_mux_set_parent and then use that new function as the
main body of clk_mux_get_parent to look up the same currently returned
value, leading to no change in functionality for existing calls.

Fixes: SYSFW-4818 #integrate-and-build

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>