pm: clk: mux: Introduce suspend and resume handlers for mux_reg
Introduce suspend and resume handlers for the clk_drv_mux_reg driver.
This is a clock mux controlled by a register, so the suspend and resume
handlers simply save and later restore the value of the mux, using the
existing parent set/get calls.

Both the suspend and resume handlers and the context data variables are
wrapped in CONFIG_LPM_CLK to avoid any code size impact when not
implementing LPM features.

Fixes: SYSFW-4820 #integrate-and-build

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>