pm: lpm: Suspend the power master in the low power entry path
The power master must be suspended during the low power entry path to
ensure proper ref counting on the relevant devices for when they are
enabled in the resume path. In a full SoC this will occur during power
master suspend sequence, but we do not have that here, so hack it for
now.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>