tisci: rm: docs update: resource range response comment fix
Fixed tisci_msg_rm_get_resource_range_resp comment for
"range_start" parameter for the documentation.
Zero is a valid value for a resource's "range_start" in the
case where range_num is non-zero.  Only when BOTH values
are zero does it indicate that the resource request is
not valid.

Signed-off-by: Jonathan Bergsagel <jbergsagel@ti.com>