pm: lpm: Deinit MAIN DEVGRP and re-init device layer
Deinit MAIN DEVGRP during the suspend path and re-init device layer
during resume. This is not the cleanest way of handling this, as this
does far more than we actually need and may in fact wipeout previous
clock non-default configuration. Additionally, it seems to take hundreds
of milliseconds. However, it works to properly reset reference counts
for now until an optimal solution is found.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>