lpm: Do not use declaration time initialization for structs
By using declaration time initialization for structs the compiler is
using a memcpy to fill in the values. This is fine during normal
operation but during low power mode this calls memcpy which sits in DDR,
which is not available, and can lead to aborts.

Avoid this by manually intializing each member of the structs.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>