lpm: Move enabling of global interrupt in dm stub
Currently the global interrupt is enabled in the entry path of the dm
stub. But because of this, if a wakeup interrupt comes before suspend
the interrupt handler will get called before entering deep sleep.
Because of this we will miss that wakeup interrupt. To fix this enable
the global interrupt after going to WFI. by this the interrupt will
wake the device from deepsleep and interrupt handler will get executed
after WFI.

Fixes: SYSFW-5026

Signed-off-by: Sebin Francis <sebin.francis@ti.com>