lpm: Add API support for dm2tifs sec proxy thread
DM to TIFS communication is done using a separate thread. Add APIs to
send and receive message on this thread.

When using the DM2TIFS queues for communication, messages must include
the tisci_sec_header. This is a 4 byte additional header prepended to
the messages. Update the driver to handle this both on send and receive,
automatically, for the DM2TIFS queue.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>
Signed-off-by: Sebin Francis <sebin.francis@ti.com>