lpm: sec_proxy: Add option for multiple sec_proxies
Allow the lpm sec_proxy driver to accept base addresses for rt and
target regions to allow use of the functionality with different
instances of secure proxy. This is required as early communication with
FS stub will happen using the ROM secure proxy and later switches to
normal TIFS configuration.

Fixes: SYSFW-4684 #integrate-and-build

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>