lpm: Enable USB reset isolation
Set USB reset isolation to active during the low power entry sequence.
Currently, we cannot yet remove this, as we must have a way to configure
USB LPSC as enabled before removing reset isolation, or the USB IP will
be reset. This is because USB LPSC comes back up out of main domain
reset in SwRstDisable state, and typically the LPSC will not be enabled
until HLOS driver requests USB LPSC again during resume.

Leaving IP in reset isolation is acceptable for now, as this will allow
USB to function as expected while also preventing reset. Only if USB is
not used at all in a system will this have impact.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>