pm: soc: am62: Sync with autogen for LPSC_sa3ul keep alive
Sync with the latest autogen to take change to add LPSC_sa3ul to
keep_alive_lpsc_list for am62 to avoid trying to shut it off, as ROM
leaves it in a state that does not allow it to be powered off, and it
will automatically be shut off in low power mode when main domain is
turned off.

If this is not done, it will cause GP_Core_CTL to report it is stuck in
transition after any attempt to power off LPSC_sa3ul, which causes many
other LSPC transitions to report timeout.

Fixes: SYSFW-5031 #integrate-and-build

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>