pm: lpm: Add delay for MAIN UART Stability
The previous patch 72232793f14c ("pm: soc: am62: Sync with autogen for
LPSC_sa3ul keep alive") allows all LPSCs to transition without falsely
reporting timeout as was occuring before, and this causes resume to
complete more quickly. A side effect of this is that garbage output can
be seen in main uart0 at the very beginning of HLOS console output, so
add a delay to prevent this.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>