am62: Autogen changes from pulling in OPTEE sproxy thread
In order to be in sync with jacinto devices, a dedicated
sproxy thread for OPTEE was added at index 11, and the old
thread at index 11 was moved to the ICSSG spot. An additional
event and VINT are reserved as a result. These are the resulting
changes pulled in from autogen

Fixes: SYSFW-5412

Signed-off-by: Dillon Frank <d-frank@ti.com>