Add tisci message to get debug trace config
Debug trace configuration information is present in the base board
config. But base board config is only given to TIFS not to DM by
the bootloader. But DM needs to know what is the debug trace
configuration present in the board config for configuring DM's
trace. Add a new tisci message to get the trace configuration
from TIFS.

Fixes: SYSFW-5366

Signed-off-by: Sebin Francis <sebin.francis@ti.com>