tisci: Add tisci message to query fw capabilities
Depending on the soc/fw the supported capabilities variers. Add a tisci
message to query fw capabilities. On the responce message a 64 bit data
is placed. each bit in the data represents one fw capability.
The first bit of the data is a generic bit and will be set for all
devices.

Fixes: SYSFW-5553

Signed-off-by: Sebin Francis <sebin.francis@ti.com>