pm: am62: am62a: Add LPSC_main_test to keep_alive_lpsc_list
Add LPSC_main_test to keep_alive_lpsc_list for am62 to avoid trying to shut
it off.

Currently the efuse controller operations leads to crash on TIFS after
pm_init. On am62x and am62ax there is a bug that associates the mastering SCR
(to EFUSE CTRL) with LPSC_main_test PWR_SCR_DISABLE and the workaround is
to keep this LPSC ON.

Ref: https://jira.itg.ti.com/browse/K3_OPEN_SI-491

Fixes: SYSFW-5619

Signed-off-by: Vishal Mahaveer <vishalm@ti.com>