lpm: Fix IO isolation bug
In the timeout implementation of IO isolation the loop count was getting
rested inside the loop, update the logic to avoid that. Inside the if
condition after doing a logic AND operation the result was compared
against wrong value, update the if condition also.

For enabling the IO isolation there is two set of registers that needs
to be updated. Update both the registers for enabling IO isolation.

Fixes: SYSFW-5870

Signed-off-by: Sebin Francis <sebin.francis@ti.com>