drivers: rat: fix range check in ioremap_internal()
While calculating rat map offset for a free mapped address range, the
given address is checked whether it falls between the higher limit
and lower limit of the free mapped region. The lower and upper address
is obtained using rat_get_free_map_addr_range() API. For example let the rat
mapped address be from 0xE000_0000 to 0xFFFF_FFFF (512MB)
rat_get_free_map_addr_range API returns 0xE000_0000 as the lower
address and 0x1_0000_0000 as the higher address. These values are used
to check whether the given address falls in between the range. During
the check the 64bit  lower and upper address is typecasted into 32 bit
integer. Because of this the highAddr 0x1_0000_0000 get truncated to
0. This leads to wrong calculation. Remove the typecast to fix
the issue.

Fixes: SYSFW-6175

Signed-off-by: Sebin Francis <sebin.francis@ti.com>