lpm: Add no optimization flag for memeset and memcpy
During LPM DDR is not available, because of this DM stub is running from TCM
memory. Since DDR is not available a separate c runtime init, stack, lpm related
functions are kept in TCM memory. But due to compiler optimization the
lpm_memset() and lpm_memcpy() are replaced with compiler provided functions.
These compiler provided functions are kept in DDR and DDR is not available
during LPM. Add (no_builtin("")) attribute flag to the function so that it will
not get optimized.

Fixes: SYSFW-6226

Signed-off-by: sebin francis <sebin.francis@ti.com>