Revert "pm: Modify PLL driver to calculate VCO frequency"
This reverts commit 4474ec516c164e47b8ea1ca75c7868ea9af898e1.

The change in PLL driver to calculate VCO frequency is causing
regression with AM65x and AM64x Linux boot. On AM65x the entire
SPL, u-boot and Linux console logs have garbled output and on AM64x
few characters in bootloader output is garbled.

Reverting the change to fix the regression.

Fixes: SYSFW-6251

Signed-off-by: Vishal Mahaveer <vishalm@ti.com>