pm: soc: am62a: Update default VCO frequency of DDR PLL
Update the default VCO frequency of DDR PLL on am62ax to match LPDDR4
configuration (1866.5 MHz) instead of DDR4 configuration (1600 MHz).

LPDDR4 is the default configuration used on AM62A, so matching the
default PLL settings also. With this approach the bootloader can just
switch in and out of bypass as part of initialization.

Fixes: SYSFW-6259

Signed-off-by: Vishal Mahaveer <vishalm@ti.com>