pm/rm/utils: fix MISRA.ETYPE.ASSIGN.2012
The value of an expression shall not be assigned
to an object with a narrower essential type or
of a different essential type category.
Fixed by having the required datatype.

sbool clk_set_parent - function is modified:
second parameter new_parent size changed to u16 -> u8,value of
new_parent get upadted by idx varible which size is 6-bit olny.

Fixes :SYSFW-6414

Signed-off-by: sangeethaduraisamy <x1096604@ti.com>