pm/pll: Update pll calculation for pll_type_16fft to use PLL table VCO options.
If the requested frequency cannot be achieved by changing the hsdiv and clk_modify_parent_frequency
is enabled for that hsdiv, the VCO frequency of the PLL will be altered to obtain the required frequency.
However, in this case, the VCO frequency can vary between 800MHz to 3.2 GHz for 16fft PLLs, and the user has
no control over it. To tackle this issue, if the required frequency is not achievable, one can refer to the PLL
table, which contains a list of alternate suggested frequency values for pllm and pllf for the corresponding PLL.

If the target frequency cannot be achieved using values from the PLL table,then use pll_internal_calc
 function to determine the optimal VCO frequency.

Fixes: SYSFW-6389

Signed-off-by: Chandru Dhavamani <chandru@ti.com>