lpm: update the delay function
The delay function inside lpm uses dummy CPU cycles for delay. Because
of this the delay varies with the frequency in which the core is
running. Update the delay logic to calculate the dummy cycles using the
CPU frequency.

Increase the PSC transition timeout since the AM62A DDR PSC is taking
time to transition.

Fixes: SYSFW-5750

Signed-off-by: Sebin Francis <sebin.francis@ti.com>