soc: am62/am62p: Reduce the DSS clocks default frequencies
Currently on AM62x and AM62Px some of the DSS PLLs/HSDIVs are
set up for higher defaults.

Since the OLDI IP is not always in use, the "by-7" clock divider between
the PLL and DSS is in bypass, and the DSS receives the high frequency clock.

Ideally, these PLLs should get set to bypass values, but since we don't
have the framework for initializing PLLs to bypass, update the defaults
to within the acceptable DSS range.

Fixes: SYSFW-6774

Signed-off-by: Vishal Mahaveer <vishalm@ti.com>