trace: Allocate sufficient bit-space for LPSC IDs
LPSC IDs can range from 0-127, taking up to 7 bits. However, only 6 bits
were allocated in the trace definition before. This creates a truncation
of the LPSC ID in trace messages and an overflow into the PSC ID.

To fix this updating the bit-space of LPSC by allocating 7 bits. Even though
PD IDs bits are not affected, updating them to be in sync.

Fixes: SYSFW-4075

Signed-off-by: Vishal Mahaveer <vishalm@ti.com>