rm: Update the IR save/restore size limit
Currently the IR save/restore functionality
has an option to either support all possible IR
configurations, or to only support a limited
number of configurations

Particularly for the limited support option,
the value has been increased to 512. This
comes from adding up the total amount of
DMA flows on AM62x/AM62a/AM62p, and multiplying
by 2, which results in 474. So, the next
highest multiple of 2 was chosen, which is 512.

Fixes: SYSFW-7103

Signed-off-by: Dillon Frank <d-frank@ti.com>