pm: soc: am62a/am62p: Add TIFS LPSC to keep alive list
On am62a/am62p devices there are no devices connected to this
LPSC and as part of pm_init the lpsc will be turned off if there
are no dependency to maintain.

When HSM related LPSC will be hidden from pm operations, it will
result in dependency graph that would result in turning off this LPSC.
Avoid it by keeping it in keep alive list.

Fixes: SYSFW-7375

Signed-off-by: Vishal Mahaveer <vishalm@ti.com>