lpm: Add mode selection logic
Mode is selected depending on three
factors.
1. Device constraints - Select mode
on basis of constrained device power
domain
2. Latency constraints - Select mode
on basis of desired wkup latency
3. Padcfg values - Select mode on basis
of wkup enabled on a particular pad.
If any of the constraints are set, the
shallowest of all these modes will
be selected.
Otherwise, the default deepest lpm mode
will be selected.

Add support to get LPSC and devgroup from
Device ID.

Fixes: SYSFW-6759

Signed-off-by: Akashdeep Kaur <a-kaur@ti.com>