pm: clock: clk_pll_16fft: Check cal lock only if cal_en is set
Currently cal_lock is checked based on pll_type and integer mode
check. There are possibilites that the calibration settings were
not done for a pll that meets these conditions.

Add an additional check based on cal_en bit to make sure calibration
settings were done on the pll that is checking for a cal_lock.

Fixes: SYSFW-7486

Signed-off-by: Vishal Mahaveer <vishalm@ti.com>