pm: clk_pll_16fft: Update the initialization sequence of HSDIVs
In current PLL initialization sequence HSDIVs are enabled and
configured before configuring the PLL and while PLL is not in
bypass. This could result in scenarios where HSDIVs are enabled
without knowing the state of PLLs.

Updating the sequence to initialize HSDIVs after the PLL is
configured so we always end up using safe combination of VCO
and HSDIVs.

Fixes: SYSFW-7486

Signed-off-by: Vishal Mahaveer <vishalm@ti.com>