pm: soc: j7: Fix MISRA.ETYPE.ASSIGN.2012
"The value of an expression shall not be assigned to an object
with a narrower essential type or of a different essential type
category"

readl and pm_writel_verified function has parameters of type u32
passing unsigned long int value results in MISRA.ETYPE.ASSIGN.2012.

Fixed this violation by changing macro values to unsigned int.

Signed-off-by: Chandru Dhavamani <chandru@ti.com>